Remap syntax highlighting function
The <C-S-P> mapping for the function to show the syntax highlighting
groups for the current cursor position collides with the AUR package
"plotinus-git" (1) which provides a searchable command palette in every
modern GTK+ application. Is uses the commonly known
"Control + Shift + P" key mapping. The used terminal emulator "tilix" (2)
is a GTK3 application so the command bar is available and has a higher
priority as the mapping for the function.

This commit changes the key mapping for the function to use the "K"
key instead of "P".

References:
  (1) https://aur.archlinux.org/packages/plotinus-git
  (2) https://github.com/gnunn1/tilix

GH-36