Add JSX syntax plugin
The "mxw/vim-jsx" (1) plugin adds syntax highlighting and indenting
support for JSX. It is not a standalone JavaScript plugin and must be
used with a base JS highlighter and integrates with the currently used
and recommended "pangloss/vim-javascript" (2) plugin (vim-jsx's
"official" dependency).

The "g:jsx_ext_required" is set to "0" to also enable JSX synta
highlighting in ".js" files.

References:
  (1) https://github.com/mxw/vim-jsx
  (2) https://github.com/pangloss/vim-javascript

GH-44