Add new "diff-so-fancy" compatible patch alias
This commit adds the new "fp" (format patch) alias to fix a output
problem when using "diff-so-fancy" with the the "format-patch" command
which also printed the ANSI color sequences. This is prevented by using
the "--no-color" option.

The alias also contains the "--patch-with-stat" option to add additional
patch stats like the total amount of changed/removed lines.

GH-84