Remove custom ATOM_HOME path
The "ATOM_HOME" environment variable (1) was assigned to a custom path
for the configuration folder of Atom instead of using the default
"~/.atom". The reason was that the folder was located in a kind of
"global application configuration collection folder" on another drive
mounted on system start. This has now been deprecated and changed to
use the default paths of the installed applications. The "~/.atom"
folder is now used for Atom which is located on the main SSD drive
instead of using a symlink to the external drive.

The change is enormous performance boost as the read/write capacity of
the external drive HDD was of course way slower than the SSD.

References:

  (1) http://flight-manual.atom.io/using-atom/sections/basic-customization/#custom-home-location-with-an-environment-variable

GH-98