Add new snowblock for Taskwarrior
Create a new snowblock for Taskwarrior (1) including

* the "key=value" based `.taskrc` configuration (2) file
* a Nord (3) based color theme to be included in `.taskrc`

> Content

>> Context

The context (4) feature should be used to separate all public projects
and private/job related tasks.

>> UDA (User Defined Attributes)

The UDA (5) feature should be used to create custom attributes to add
a assignee, a estimation value and the GitHub issue ID to a task.

>> Reports

The reports (6) feature should be used to extend and optimize the task
output. The custom report should be set to `default.command` to replace
the default `next` report.

>> Aliases

The config should contain aliases to align with the alias workflow of
other snowblocks like bash and git. This improves the usability,
because there are several commands that start with the same characters
making it slower to be used with the auto completion.

* `a` ->  `add`
* `bd` ->  `burndown.daily`
* `cx` ->  `context`
* `e` ->  `edit`
* `ls` ->  `list`
* `mod` ->  `modify`

> Alternative workflows

The Taskwarrior workflow could be changed later on to use multiple
databases per use-case like e.g. one database per project to make
better use of the features like projects, tags and context with
finer grained filters.

> Configuration

>> Core Plugins

>>> `clean`

* `~`
* `~/.task`

>>> `link`

* "archbook":
  * `taskrc` -> `~/.taskrc`
  * `nord.theme` -> `~/.task/nord.theme`
* "igloo":
  * `taskrc` -> `~/.taskrc`
  * `nord.theme` -> `~/.task/nord.theme`

References:

  (3) https://arcticicestudio.github.io/nord
  (1) https://taskwarrior.org
  (2) https://taskwarrior.org/docs/introduction.html
  (4) https://taskwarrior.org/docs/context.html
  (5) https://taskwarrior.org/docs/udas.html
  (6) https://taskwarrior.org/docs/report.html

Closes GH-107