Add time tracking hooks
Implemented two new hooks to track the duration a task was active. Both
hooks are written in Python 3 and require a Taskwarrior version `>=2.4`.

> With `totalactivetime` UDA

The first hook makes use of a new UDA called `totalactivetime` of type
`duration` which stores the total time in seconds. Every time a task
gets stopped the value gets added to the old value. To simplify the
communication with the Taskwarrior Hook v2 API (1) the hook should make
use of the "taskw" (2) package which provides the bindings for Python.

NOTE: This requires "taskw" (2) as a direct dependency and must be
installed on the target system!
The package is available from the AUR (3).

> Taskwarrior integration

The second hook should be a fort from the official Taskwarrior
`on-modify.taskwarrior` hook (4). It comes bundled with Taskwarrior,
but can not be used via symlink, because the file is owned by the
`root` user which prevents the script from being executed when symlinked.
The usage of the hook is also documented in the official Timewarrior docs (5).

References:

  (1) https://taskwarrior.org/docs/hooks2.html
  (2) https://pypi.python.org/pypi/taskw
  (3) https://aur.archlinux.org/packages/python-taskw
  (4) https://github.com/GothenburgBitFactory/timewarrior/blob/dev/ext/on-modify.timewarrior
  (5) https://taskwarrior.org/docs/timewarrior/taskwarrior.html

Closes GH-108