Set color profile to sRGB
Atom 1.28 introduced support for custom ICC color profiles (1) based on
the upgrade to Electron 2.x which in turn is based on the latest Chrome
version. This version includes changes to the color rendering and now
respects the color profile configured in the OS settings.

Due to this change for many users the [colors appear to be pale (2).
To bring back the previous (more stronger) color rendering
Atoms „Core“ → „Color Profile“ must be changed from „Use color profile
configured in the operating system“ to „Use sRGB color profile“.

 References:

   (1) http://blog.atom.io/2018/06/21/atom-1-28.html#support-for-custom-color-profiles
   (2) https://github.com/atom/atom/issues/17356

Closes GH-115