Allow more than active task
The default Taskwarrior limit of active tasks is one. However, there are
many use cases where multiple active tasks must be used, e.g. when
working on an GitHub issue and the associated PR which are both handled
in separate tasks or when working on tasks that are dependent on other
tasks.
The limited has been increased to allow to use Taskwarrior for such use
cases without having to deal with limitation errors.

Closes GH-117