Add package configuration and alias for Buku
Buku (1) provides the `BUKU_COLORS` environment variable to allow users
to define a custom color scheme. It follows the same principle like the
`LSCOLORS` environment variables of the BSD `ls` command.
The five letters represent the colors in ordered form of

* index
* title
* url
* description/comment/note
* tag

See the GitHub repository wiki page about how to customize colors (2)
for details about available letters and their meaning.

Next to this a shortcut alias `bk` should be added to simplify the usage
commands.

References:

  (1) https://github.com/jarun/Buku
  (2) https://github.com/jarun/Buku/wiki/Customize-colors

Closes GH-128