Add macOS support for GPG
> Changed the default socket file paths.

There was a problem where the `gpg-agent` was not able to create the
socket files within the `~/.gnupg` folder that is symlinked to the
gocryptfs (1) container to securely (encrypted) store the GPG keys:

```
gpg: can’t connect to the agent: IPC connect call failed
```

This might be caused by either the macOS filesystem (APFS) that doesn't
support the creation of sockets (it works fine on other Arch Linux hosts
like `igloo`) or by restrictions due to read/write permissions that are
not passed through.

This solution/workaround is to change the default absolute file paths of
the socket files:

1. Created a `S.gpg-agent` and `S.gpg-agent.ssh` file manually in the
   `$GNUPGPHOME` folder (defaults to `~/.gnupg`).
2. Added the `extra-socket` and `browser-socket` options in the
   `gpg-agent.conf` file with the desired paths.

The created `S.gpg-agent` and `S.gpg-agent.ssh` files allow to use
values of environment variables via string interpolation (2), e.g.
`${HOME}`.

NOTE: It is important to make sure that the target folder exists and
the permissions are set to 700!

> macOS pinentry tool

Set the `pinentry-programm` option in the `gpg-agent.conf` file to use
pinentry-mac (3). Note that this requires the Homebrew formula
pinentry-mac (4) to be installed!

```conf
pinentry-program /usr/local/bin/pinentry-mac
```

> More references

* „gpg: can’t connect to the agent: IPC connect call failed“ (5)

* GNUPG bugtracker - “gpg-agent 2.1 socket and nfs /home“ (6)
* „How to configure GnuPG's S.gpg-agent socket location?“ (7)

References:

  (1) https://github.com/rfjakob/gocryptfs
  (2) https://en.wikipedia.org/wiki/String_interpolation
  (3) https://github.com/GPGTools/pinentry-mac
  (4) https://formulae.brew.sh/formula/pinentry-mac
  (5) https://michaelheap.com/gpg-cant-connect-to-the-agent-ipc-connect-call-failed
  (6) https://dev.gnupg.org/T1752
  (7) https://askubuntu.com/a/1053594

Epic: GH-131
Closes GH-132