taskopen workaround support for macOS
The management of installed Perl modules (1) on macOS is not as simple
and well thought through like with package managers on Linux systems,
e.g. via pacman (2) on Arch Linux (3). There are problems when
is comes to configuring the runtime path the modules have been installed
to even when using the most popular module manager called cpanminus (4).
This causes the Taskwarrior (5) plugin taskopen (6) fail to load because
the Perl core module `JSON` can't be found and loaded.

As a workaround a custom script has been implemented to create and open
a attached note of an task:

1. Used the `_get` (7) function of the Taskwarrior DOM
   API (8) to extract any stored piece of information of an task. This
   allows to receive the UUID of an task (9).
2. Created a custom `on` (open note) Taskwarrior alias (10) to run the
   implemented custom script via the `execute` command.

The logic of the script follows the same like taskopen uses for default
notes:

* Using the UUID of an task (TODO) as the note filename.
* Simply passing the file to an editor (in this case Atom (11)) which
  will…
    * …creates a new file if it doesn't exist yet.
    * …opens the file if it already exists.

Note that this script is not limited to macOS but can also be used for
any other Linux hosts!

References:

  (1) http://www.cpan.org/modules
  (2) https://wiki.archlinux.org/index.php/Pacman
  (3) https://archlinux.org
  (4) https://github.com/miyagawa/cpanminus
  (5) https://taskwarrior.org
  (6) https://github.com/ValiValpas/taskopen
  (7) https://taskwarrior.org/docs/commands/_get.html
  (8) https://taskwarrior.org/docs/dom.html
  (9) https://taskwarrior.org/docs/ids.html
  (10) https://taskwarrior.org/docs/terminology.html#alias
  (11) https://atom.io

Epic: #131
Related to #110
Closes GH-134