JetBrains IDE configurations snowblock
This new snowblock includes all important configurations for all used
JetBrains (1) IDEs and products which are currently IntelliJ (2)
(Ultimate Edition), PyCharm (3) (Community Edition), Webstorm (4) and
GoLand (5). For details about the paths see the documentation about
directories used by the IDE to store settings, caches, plugins and logs
(6).

It includes…

* …the code style (7) XMl configuration file stored in the `codestyles`
  folder.
* …the inspections (8) XMl configuration file stored in the
  `inspections`  folder.
* …the keymaps (9) XMl configuration files for Linux and macOS stored in
  the `keymaps` folder. See the keymap reference (10) for a visual
  presentation of the default mappings.
* …the preferences (11) XMl configuration files stored in the `options`
  folder. This includes the following settings:
  * `code.style.schemes.xml` - settings of the currently used code style.
  * `colors.scheme.xml` - settings of the currently used color scheme.
  * `customization.xml` and `project.default.xml` - settings of the
    customized menus and toolbars (12).
  * `editor.codeinsight.xml` - settings of various automatic code
    transformations.
  * `keymap.xml` - the currently used keymap.
  * `laf.xml` and `ui.lnf.yml` - settings of the currently used UI theme.
  * all other customized preferences defined in the various categories:
  `editor.xml`, `ide.general.xml` and `options.xml`
* …the list of disabled plugins (13) stored in the
  `disabled_plugins.txt` file.

>> snowsaw core plugin configuration

NOTE: JetBrains uses a naming scheme for the folder all user
configurations are stored in based on the product major/minor version!
This requires this snowblock to be updated as soon as the product gets
updated to adapt to the new version string.

>>>>>> `clean`

The target paths (based on the host OS) with the naming scheme
`<PRODUCT><VERSION>` should be cleaned before the `link` core plugin
execution.

NOTE: Due to the JetBrains naming scheme for the target path (more
details described above) the previous targets should be kept to ensure
they are cleaned up by this plugin after updating to a new product
version.

>>>>>> `link`

The target paths (based on the host OS) with the naming scheme
`<PRODUCT><VERSION>` should be created by snowsaw if they doesn't exist
yet.

NOT: Due to the JetBrains naming scheme for the target path (more
details described above) must be updated after updating to a new product
version.

References:

  (1) https://www.jetbrains.com/products.html
  (2) https://www.jetbrains.com/idea
  (3) https://www.jetbrains.com/pycharm
  (4) https://www.jetbrains.com/webstorm
  (5) https://www.jetbrains.com/go
  (6) https://intellij-support.jetbrains.com/hc/en-us/articles/206544519-Directories-used-by-the-IDE-to-store-settings-caches-plugins-and-logs
  (7) https://www.jetbrains.com/help/idea/settings-code-style.html
  (8) https://www.jetbrains.com/help/idea/inspections-settings.html
  (9) https://www.jetbrains.com/help/idea/settings-keymap.html
  (10) https://www.jetbrains.com/help/idea/keymap-reference.html
  (11) https://www.jetbrains.com/help/idea/settings-preferences-dialog.html
  (12) https://www.jetbrains.com/help/idea/menus-and-toolbars.html
  (13) https://www.jetbrains.com/help/idea/managing-plugins.html

Closes GH-137