Replace/Update "archbook" with new "iceowl" host
There are various snowblock/snowflake configurations that contain the
`archbook` host (1) which doesn't exist anymore and can be replaced in
most cases with the new `iceowl` host. In cases where the
snowblock/snowflake is only targeted for Linux the `archbook` host
should be removed to make it host independently. These can be updated
when a new Linux host is introduced.

References:

  (1) https://github.com/search?l=&q=archbook+repo%3Aarcticicestudio%2Figloo+created%3A%3E2018-09-18&type=Code

Epic: GH-131
Closes GH-143