Add alias for IINA
A new `iina` alias for IINA (1) should be added to ensure the MPV (2)
specific flags `--mpv-config-dir` (3) and `--mpv-watch-later-directory`
are always used which contain the paths to the MPV configuration
files (4) and the "watch-later" folder (5) of the
`--save-position-on-quit` feature (6). This ensures the "watch-later"
state/position files are always synced with `igloosync` to ensure
consitency with other synced hosts like `igloo`.

```sh
iina --mpv-config-dir=~/.config/mpv --mpv-watch-later-directory=~/.config/mpv/watch_later
```

NOTE: This alias is only for `macOS`!

References:

  (1) https://lhc70000.github.io/iina
  (2) https://mpv.io
  (3) https://mpv.io/manual/master/#options-config-dir
  (4) https://mpv.io/manual/master/#files-~/-config/mpv/mpv-conf
  (5) https://mpv.io/manual/master/#files-~/-config/mpv/watch-later
  (6) https://mpv.io/manual/master/#options-save-position-on-quit

Closes GH-147