Disable "Find As You Type"
The "Find As You Type" (1) feature (UI path: Preferences → General →
Browsing → "Search for text when you start typing") should be disabled.
It allows to search the current site as soon as a key is pressed instead
of requiring to press "Ctrl"/"⌘" + "f". This is a great accessibility
improvement, but unfortunately this often collides with page features
like hotkeys e.g. when pressing "r" on Octobox (2) to sync the latest
notifications or any other site using such features.

References:

  (1) http://kb.mozillazine.org/Accessibility.typeaheadfind
  (2) https://octobox.io

Closes GH-164