Implement shorthand script to open files, paths or URLs
Most OS provide builtin tools to open a file, path or URL with the
associated application based on the MIME type. To use a uniform command
for it this commit adds a new script called `open`.
When running on Linux it'll use the XDG compliant tool `xdg-open` from
xdg-utils (1) while it'll pass through the parameters to the builtin
application also called `open` when run on macOS.

References:
  (1) https://www.freedesktop.org/wiki/Software/xdg-utils

Closes GH-169