Visual Studio Code
This commit includes the migration from Atom (1) to
Visual Studio Code (2).
I've used Atom since it came out some years ago and it has always been a
great editor and IDE for web development as well as main editor for
documents, but VS Code evolved a lot during the last year.

Compared to Atom there are many features and improvements that I've
really liked to see for Atom. Even though it can be extended and
modified a lot there are limitations due to Atom's core in aspects like
performance, stability and overall development features like auto
completion and language-awareness.

VS Code feels more "language-native" when it comes to e.g. intelligent
auto-completion, the integration of linters and language specific tools.
It comes with support for a lot of languages out-of-the-box like
JavaScript/TypeScript, CSS/HTML and also native languages like Go or
Rust.

The change was not easy for me since I've customized and developed Atom
to fit all of development needs, but after some testing time I've to
admit that the UX of VS Code feels better now.

Before actually switching I made sure that all my beloved Atom packages
are also available as VS Code extension (same code base) or an
equivalent alternative.

Therefore this initial setup includes

- core and extension settings
- custom keybindings
- locale configurations
- a list of all installed extensions
- CSS stylesheets to customize some that'll be loaded by the "Custom
  CSS and JS Loader" (3) extension
- workspace files

The following extensions are used:

- Nord (4)
- file-icons (5)
- Prettier (6)
- ESLint (7)
- Markdown Preview Enhanced (8)
- Babel JavaScript (9)
- vscode-styled-components (10)
- Path Intellisense (11)
- Go (12)
- Docker (13)
- YAML (14)
- EditorConfig (15)
- Projects+ (16)
- Polacode (17)
- Custom CSS and JS Loader (18)
- Rust (rls) (19)
- GraphQL (20)

References:
  (1) https://atom.io
  (2) https://code.visualstudio.com
  (3) https://marketplace.visualstudio.com/items?itemName=be5invis.vscode-custom-css
  (4) https://marketplace.visualstudio.com/items?itemName=arcticicestudio.nord-visual-studio-code
  (5) https://marketplace.visualstudio.com/items?itemName=file-icons.file-icons
  (6) https://marketplace.visualstudio.com/items?itemName=esbenp.prettier-vscode
  (7) https://marketplace.visualstudio.com/items?itemName=dbaeumer.vscode-eslint
  (8) https://marketplace.visualstudio.com/items?itemName=shd101wyy.markdown-preview-enhanced
  (9) https://marketplace.visualstudio.com/items?itemName=mgmcdermott.vscode-language-babel
  (10) https://marketplace.visualstudio.com/items?itemName=jpoissonnier.vscode-styled-components
  (11) https://marketplace.visualstudio.com/items?itemName=christian-kohler.path-intellisense
  (12) https://marketplace.visualstudio.com/items?itemName=ms-vscode.Go
  (13) https://marketplace.visualstudio.com/items?itemName=PeterJausovec.vscode-docker
  (14) https://marketplace.visualstudio.com/items?itemName=redhat.vscode-yaml
  (15) https://marketplace.visualstudio.com/items?itemName=EditorConfig.EditorConfig
  (16) https://marketplace.visualstudio.com/items?itemName=fabiospampinato.vscode-projects-plus
  (17) https://marketplace.visualstudio.com/items?itemName=pnp.polacode
  (18) https://marketplace.visualstudio.com/items?itemName=be5invis.vscode-custom-css
  (19) https://marketplace.visualstudio.com/items?itemName=rust-lang.rust
  (20) https://marketplace.visualstudio.com/items?itemName=Prisma.vscode-graphql

Resolves GH-179