Basic Go environment setup
I started to use Go (1) as one of my main languages for many cases like
CLI tools, backends with performance and inter compatibility in mind
as well as daily tools.
Go comes with a great toolchain that also doesn't require a lot of
necessary setups, therefore this basic setup is enough to work
efficiently in combination with the awesome IntelliJ Go plugin (2) and
the great integration of Go in VS Code (3).

This basic setup includes:

- Setting the `GOPATH` envionment variable to the custom location at
  `$HOME/.go`. Even though that I only use Go Modules (4) this is
  required for backwards compatibility, the toolchain as well as the
  resulting binaries in `$GOPATH/bin`.
- Add `$GOPATH/bin` to `$PATH` for all hosts.

References:
  (1) https://golang.org
  (2) https://plugins.jetbrains.com/plugin/9568-go
  (3) https://code.visualstudio.com/docs/languages/go
  (4) https://github.com/golang/go/wiki/Modules

Resolves GH-181