Custom KeePassXC desktop launcher
KeePassXC (1) supports GTK3 out-of-the-box, but unfortunately doesn't
adapt to the current system theme and icon them making the UI looking
really unpleasant and outdated.
To use a modern style the app is now started with the
`QT_QPA_PLATFORMTHEME` environment variable (2) set to `gtk2` to ensure the
app respects the Qt5 style overrides (3).

References:
  (1) https://keepassxc.org
  (2) https://wiki.archlinux.org/index.php/qt#All_Qt5-based_applications_fail_to_run_after_Qt5_update
  (3) https://github.com/keepassxreboot/keepassxc/issues/1931

Resolves GH-204