Fix invalid JetBrains symlink target parent paths on Linux hosts
On macOS the JetBrains configuration files were placed right into the
main `IntelliJIdea2019.2` directory while on Linux based systems the
folder consists of the sub-folders `cache`, `config` and `system` that
can also be configured in the `idea64.properties` file.
Previously the target path was also the main `~/.IntellihJIdea2019.2`
directory, but the files must be linked into the nested `config` folder
instead.

Fixes GH-212