Incremental naming of copied files/folders
In VS Code version 1.37 (1) the automatic incremental naming for copied
files that include numbers has been (silently) removed, but than added
back as explorer improvement in version 1.38 (2) with two configuration
options to either add a `copy` suffix or use a `smart` strategy.
The second one adds a number at the end and if some number is already
part of the name, it tries to increase that number. This reflects the
behavior from VS Code versions `<1.38` and is therefore now used.

References:
  (1) https://code.visualstudio.com/updates/v1_37
  (2) https://code.visualstudio.com/updates/v1_38#_explorer-improvements

Resolves GH-215