Add “Inter“ as base sans-serif typeface
Before it was required to manually configure the sans-serif typeface
based on the running OS [1] due to licensing and the availability of
specific font families on different operating systems.
Since I use the fantastic Inter [2] typeface family for almost all UI,
text and designs it is now set as default sans-serif typeface.
Inter is open source and therefore system independent and is installed
in all of my systems/habitats.

[1]: https://github.com/arcticicestudio/igloo/tree/eb222438e6cd93b19c222043570bb7c627ea3e3b/snowblocks/firefox#typeface
[2]: https://rsms.me/inter

GH-225