Disable password generation suggestion for password fields
Firefox comes with generally great security features like "Firefox
Lockwise" [1], but the suggestion to generate a new random password when
focusing a password field [2] is kind of annoying and not necessary due
to the usage of a password manager that includes a password generator.
Therefore it has now been disabled.

[1]: https://www.mozilla.org/firefox/lockwise
[2]: https://wiki.mozilla.org/Toolkit:Password_Manager/Password_Generation

GH-227