"cower" (+ "pacaur" configurations) snowblock decommission
cower [1] (superseded/replaced by auracle [2]) was introduced in GH-65
to configure the usage through pacaur [3] which was configured in GH-14.
`pacaur` was used since it was the cleanest AUR [4] helper from a list
of 20+ projects at the time, but some years ago finally the awesome
yay [5] project written in 💙 Go [6] was started that is not just a
shell/Python etc. wrapper around `pacman` but a great CLI + Go package
that is based on the github.com/Jguer/go-alpm [7], a Go package with
bindings to `libalpm` [8]. It also makes use of other fantastic Go
packages like github.com/Morganamilo/go-pacmanconf [9] and
github.com/Morganamilo/go-srcinfo [10o] to ensure to reduce the
requirement to "shell out" to `pacman`/`makepkg` as minimal as possible
and use pure Go code and logic instead.
This allowed me to also use and extend my `pacman` and AUR workflow on
top of my absolute favorite and main language Go.

Since pacaur has been discontinued by the original author some years
ago, which used cower [1] (GitHub repository deleted and superseded by
auracle [2]), this step was necessary anyway and I've switched to yay
almost 2 years ago.
Therefore the "cower" snowblock (and related pacaur configurations) are
not required anymore and have been removed while the data is still
available through the Git repository history/logs [11].

[1]: https://github.com/falconindy/cower
[2]: https://aur.archlinux.org/packages/auracle-git
[3]: https://aur.archlinux.org/packages/pacaur
[4]: https://aur.archlinux.org
[5]: https://github.com/Jguer/yay
[6]: https://go.dev
[7]: https://github.com/Jguer/go-alpm
[8]: https://www.archlinux.org/pacman/libalpm.3.html
[9]: https://github.com/Morganamilo/go-pacmanconf
[10]: https://github.com/Morganamilo/go-srcinfo
[11]: https://git-scm.com/book/en/v2/Git-Basics-Viewing-the-Commit-History

GH-257
Related to GH-248

Co-authored-by: Sven Greb <development@svengreb.de>