Create new snowblock for "bat"
Created a new snowblock for bat [1]
> A _cat(1)_ clone with syntax highlighting and Git integration.

I've used the project for quite some time and the configurations have
now been tested enough on a day-by-day basis to be persisted in igloo.

Initially this includes the `config` file [2] symlinked to the
corresponding XDG specific [3] and OS dependent path(s):

- Linux: `${XDG_CONFIG_HOME:-$HOME/.config}/bat`
- macOS: `${XDG_CONFIG_HOME:-$HOME/.config}/bat`

For now I'll stick with the bundled "Nord" highlighting theme,
but as soon as an official port project is available it'll either be
included in igloo or, even better, submitted to bat in order to ship it
as officially maintained theme.

[1]: https://github.com/sharkdp/bat
[2]: https://github.com/sharkdp/bat#configuration-file
[3]: https://specifications.freedesktop.org/basedir-spec/basedir-spec-latest.html

GH-259

Co-authored-by: Sven Greb <development@svengreb.de>