Create new snowblock for "broot"
Created a new snowblock for broot [1]
> A new way to see and navigate directory trees.

I've used the project for quite some time and the configurations have
now been tested enough on a day-by-day basis to be persisted in igloo.

Initially this includes the `conf.toml` file [2] symlinked to the
corresponding XDG specific [3] and OS dependent path(s):

- Linux: `${XDG_CONFIG_HOME:-$HOME/.config}/broot`
- macOS: `${XDG_CONFIG_HOME:-$HOME/Library/Preferences}/org.dystroy.broot`

The colors will be adjusted to match the "Nord" theme [4],
but as soon as an official port project is available it'll either be
included in igloo and submitted to broot in order to include it in their
documentations as officially maintained theme.

[1]: https://github.com/Canop/broot
[2]: https://dystroy.org/broot/documentation/configuration
[3]: https://specifications.freedesktop.org/basedir-spec/basedir-spec-latest.html
[4]: https://www.nordtheme.com

GH-260

Co-authored-by: Sven Greb <development@svengreb.de>