Create new snowblock for "procs"
Created a new snowblock for procs [1]
> A modern replacement for `ps` written in Rust.

I've used the project for quite some time and the configurations have
now been tested enough on a day-by-day basis to be persisted in igloo.

Initially this includes the `config.toml` file [2] symlinked to the
corresponding XDG specific [3] and OS dependent path(s):

- Linux: `${XDG_CONFIG_HOME:-$HOME/.config}/procs`
- macOS: `${XDG_CONFIG_HOME:-$HOME/Library/Preferences}/com.github.dalance.procs`

The colors will be adjusted to match the "Nord" theme [4],
but as soon as an official port project is available it'll either be
included in igloo and submitted to procs in order to include it in their
documentations as officially maintained theme.
Note that the "Nord" color palettes must be provided through the
terminal as the 16 ANSI base colors.

[1]: https://github.com/dalance/procs
[2]: https://github.com/dalance/procs#configuration
[3]: https://specifications.freedesktop.org/basedir-spec/basedir-spec-latest.html
[4]: https://www.nordtheme.com

GH-261

Co-authored-by: Sven Greb <development@svengreb.de>