Create new snowblock for "ripgrep"
Created a new snowblock for ripgrep [1]
> A line-oriented search tool that recursively searches your current
> directory for a regex pattern.
> A modern alternative for grep.

I've used the project for quite some time and the configurations have
now been tested enough on a day-by-day basis to be persisted in igloo.

Initially this includes the `config` file [2] symlinked to the
corresponding XDG specific [3] and OS dependent path(s):

`ripgrep` doesn't look up any specific directories by default but only
checks if the `$RIPGREP_CONFIG_PATH` environment variable is set and if
so tries to load the file from the path that is set as value.

The colors will be adjusted to match the "Nord" theme [4],
but as soon as an official port project is available it'll either be
included in igloo and submitted to ripgrep in order to include it in their
documentations as officially maintained theme.
Note that the "Nord" color palettes must be provided through the
terminal as the 16 ANSI base colors. The colors could also be specified
as HEX tripplets, but this would break the highlighting for terminals
that don't support "true colors", but the port project should provide
both colorspaces later on.

[1]: https://github.com/BurntSushi/ripgrep
[2]: https://github.com/BurntSushi/ripgrep/blob/master/GUIDE.md#configuration-file
[3]: https://specifications.freedesktop.org/basedir-spec/basedir-spec-latest.html
[4]: https://www.nordtheme.com

GH-262

Co-authored-by: Sven Greb <development@svengreb.de>