ISO-8601 compliant time style format for `date` command
Before there was a unnecessary duplicate colon in the `TIME_STYLE`
environment variable causing a invalid and non-parsable output.
In order to fix this and ensure a ISO-8601 [1] compliant format the
superfluous colon has been removed and the space between the date and
time replaced with a `T` character like documented in the specification.

[1]: https://en.wikipedia.org/wiki/ISO_8601
[2]: https://github.com/arcticicestudio/igloo/blob/18b735009f2baa29e3bbe1323a1fc2082f88b393/snowblocks/zsh/pkgs/coreutils/env.zsh#L10

GH-277

Co-authored-by: Sven Greb <development@svengreb.de>